requirements.txt
opencv-contrib-python
numpy
pillow
pandas
openpyxl
